workers/thumbnail/requirements.txt
opencv-python-headless
numpy
boto3
psycopg2-binary
redis
Pillow
